backend/requirements.txt
flask
flask-cors
flask-limiter
flask-sqlalchemy
pandas
numpy
scikit-learn
pyjwt
bcrypt
python-dotenv
gunicorn
scipy
pymongo
joblib
pytest
sqlalchemy
psycopg[binary]
redis
